abstracted utils functions to a new module
