+++ main.py
@@ -2,6 +2,7 @@ import os
 from dotenv import load_dotenv
 from pymongo import MongoClient
 from sqlalchemy import create_engine
+from utils import validate_mongo_connection, upload_data_to_neon
 import pandas as pd
 
 load_dotenv()
@@ -16,30 +17,6 @@ mongo_connection_string = f"mongodb+srv://{os.getenv('MONGO_USERNAME')}:{os.gete
 mongo_client = MongoClient(mongo_connection_string)
 mongo_database = "Lab_One_10"
 db = mongo_client[mongo_database]
-
-
-# logic for data movement
-def upload_data_to_neon(dataframe):
-    try:
-        dataframe.to_sql("my_table", con=engine, index=False)
-    except Exception as e:
-        print(f"An error occurred while uploading data to Neon: {e}")
-
-
-def download_data_from_neon():
-    pass
-
-
-def upload_data_to_mongo(data):
-    pass
-
-
-def validate_mongo_connection():
-    try:
-        mongo_client.admin.command("ping")
-        print("Pinged your deployment. You successfully connected to MongoDB!")
-    except Exception as e:
-        print(e)
 
 
 def parse_parquet(file):
